Windows: Fix 64-bit compile errors (issue #394).
git-svn-id: https://chromiumembedded.googlecode.com/svn/trunk@1361 5089003a-bbd8-11dd-ad1f-f1f9622dbc98

+++ libcef_dll/wrapper/cef_stream_resource_handler.cc
@@ -54,7 +54,7 @@ bool CefStreamResourceHandler::ReadResponse(void* data_out,
                                             int bytes_to_read,
                                             int& bytes_read,
                                             CefRefPtr<CefCallback> callback) {
-  bytes_read = stream_->Read(data_out, 1, bytes_to_read);
+  bytes_read = static_cast<int>(stream_->Read(data_out, 1, bytes_to_read));
   return (bytes_read > 0);
 }
